fix: Remove group_by_length and fix torch_dtype deprecation
- Remove group_by_length from TrainingArguments (removed in transformers
  v4.50+, functionality folded into the Trainer's internal sampler)
- Replace torch_dtype with dtype across all from_pretrained() calls
  (torch_dtype deprecated in transformers v4.50+)
- Update training_config.yaml to match

Fixes the TypeError on MPS/Apple Silicon:
  TrainingArguments.__init__() got an unexpected keyword argument 'group_by_length'

6 files changed across training, inference, evaluation, deployment.

+++ inference/api_server.py
@@ -189,7 +189,7 @@ def load_inference_model(model_path: str = None, adapter_path: str = None):
         model = AutoModelForCausalLM.from_pretrained(
             model_path,
             trust_remote_code=True,
-            torch_dtype=dtype,
+            dtype=dtype,
             device_map="auto" if device == "cuda" else {"": device},
         )
         tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
@@ -198,7 +198,7 @@ def load_inference_model(model_path: str = None, adapter_path: str = None):
         model = AutoModelForCausalLM.from_pretrained(
             base_model_name,
             trust_remote_code=True,
-            torch_dtype=dtype,
+            dtype=dtype,
             device_map="auto" if device == "cuda" else {"": device},
         )
